Improve handling template parameter aliases
Full uses of aliased template parameters are now handled in
`IwyuBaseAstVisitor::VisitTypedefType`, in the common manner with
handling type aliases outside template instantiations. This handler only
reports type alias underlying types. The uses of the type alias
declarations themselves are reported from `IwyuAstConsumer`
implementation, as before, and not during implicitly instantiated
template traversal, because they should be already reported at
the traversal of the template as written.

`ReportTypeUse` function reports not only direct aliases of template
parameters but also when the aliased template parameter was sugared.
A corresponding test case is added.