Improve printing of driver jobs
Jobs/commands are printed as part of diagnostics and logging.

Relying on JobList::Print is the 'normal' way of doing this, but it has
a few downsides:

* The API is not stream-friendly, so combining it with other
  logging/printing requires a bit of noisy code
* It doesn't print metadata about jobs, only a shell-like command
  sequence

The second bullet might be a design goal, but since we only use this
output for debugging, we prefer to inject some more information.

Add a stream operator for Command, and prefix the command line by action
class name.

Reimplement JobList::Print as JobsToString, which uses said stream
operator to build a custom-separated string of all contained jobs.

Use these two mechanisms for all driver diagnostics.