Filter out jobs from unexpected tools in driver
Before this, we'd fail the compilation if the first job happened to be
created by a tool other than "clang".

Soften this, and instead filter out all jobs not created by "clang". If
no jobs remain after filtering, fail with 'expected exactly one compiler
job' as usual.

Also add a diagnostic at verbosity level 2 that job was dropped, so it's
possible to debug.

I don't know a way to spell a command so the jobs are created by
anything but "clang", thus no testcase.