Use CompilerInstance to create DiagnosticsEngine
It turns out CompilerInstance has a static factory method for
DiagnosticsEngine, which configures it just the way we like it.

Replace custom construction with this frontend helper.

I initially wanted to reuse the same diagnostics engine for the
compiler, but that turns out not to work -- CompilerInvocation sets
various diagnostic options based on the arguments, IWYU always runs with
defaults. Clarify with a comment why future me shouldn't try it again.

Remove redundant hasDiagnostics() check; createDiagnostics() can't fail.