Make blank lines around #pragma push/pop lines prettier
Before we would transform from this:

    #pragma GCC diagnostic push
    #pragma GCC diagnostic ignored "-Warray-bounds"
    #include <cstdint>
    #pragma GCC diagnostic pop

to this:

    #pragma GCC diagnostic push
    #pragma GCC diagnostic ignored "-Warray-bounds"
    #include <cstdint>

    #pragma GCC diagnostic pop

because we didn't recognize push/pop pragmas as special. Usually they
mark a group, and it's nicer to keep the trailing pop together with the
line before.

Fix this, and also put a blank line before the push, so as to highlight
the coming group.

The push/pop pragmas are loosely pattern matched, so the same rule will
apply to similar pragmas like:

- #pragma pack(push)/#pragma pack(pop)
- #pragma warning(push)/#pragma warning(pop)
- #pragma push_macro/#pragma pop_macro

Fixes issue #1237.