Add mappings statistics and dump
We have in-flight logging when mappings are added, but no summary
information about what the effective mappings are once IWYU analysis is
about to start.

Add high-level statistics (count/min/max/average) to help understand the
shape of the data, and also a full dump of include and symbol mappings.

These logs are designed to help guide mapping improvements and
debugging.