Avoid separate handling of fn refs from CallExprs
Previously, it was just redundant. But after the previous commit, it
started causing a bug with unwanted default type template argument
reporting when it is actually overridden by the deduced one.

The problem arises because there is a lack of the full context for
`resugar_map` and provided type set determination: actual function call
arguments are not considered on function reference analyzing. Hence,
a default type argument of a function template may be placed into
`resugar_map` as defaulted despite being overridden by some of function
call argument types.