Collect all #includes at the top of iwyu.cc
For some reason, iwyu.cc has had a separate little world around main()
-- presumably because it can't be a part of namespace
include_what_you_use, but it's hard to say.

Move the #includes to the top among the rest of them, and move the using
declarations into main().

In the process, remove a duplicate #include of FrontendAction.h, and
sort the using decls.

No functional change.