[ci] Add CI steps for IWYU dogfooding
Add a script iwyu-dogfood.bash, that runs IWYU on itself and processes
the output into a nice Markdown report called iwyu-dogfood.md.

Enable CMAKE_EXPORT_COMPILE_COMMANDS to make the build system generate a
compile_commands.json for IWYU.

Upload the report + a file containing the current PR API URL as an
artifact, to be picked up by a triggered workflow_run, which in turn
translates the file into a PR comment.

This gives us a nice summary comment on every PR showing how
IWYU-on-IWYU behaves with the PR applied, and also if the code changes
follow IWYU policy.