[ci] Fix escaping for dogfood comment curl command
The last attempt failed with:

>  "message": "Problems parsing JSON",

which I am guessing was due to double double-quoting around the
content.

The 'jq -R -s' command returns a quoted string, so we should be able to
inline it like this and get valid JSON. Local experiments with echo
indicate that it *should* work.

Fingers crossed.