[ci] Generalize PR comment handling
Turn things around so we always run the IWYU dogfood step, even for
scheduled builds. Put the step summary back.

For PRs, bundle dogfood results into pr-comments artifact to be picked
up by pr-comments-bot workflow.

Generalize wording and fix a typo in pr-comments-bot.yml.

This makes it possible to add a new informational step by simply
producing a Markdown doc in ci.yml, bundle it, and push it as a comment
in pr-comments-bot.yml.