Reject attempts at using PCH
IWYU as-is cannot support PCH, because we care deeply about the include
graph, and it does not appear to be preserved when Clang interprets the
precompiled header format.

We used to fail mysteriously with segmentation faults due to missing
files; now we fail with an explicit message on startup.

Remove null-check for this scenario in iwyu:VisitUsingDecl, that code
should now be unreachable.