Add and use more complete functionality for special files
The IsBuiltinFile and IsBuiltinOrCommandLineFile predicates were used in
a few places to determine if includers/includees were actual files.

This would fail to match some non-files like "<scratch space>", but was
also not very ergonomic, as callers would have to resolve file entry in
order to check for builtin-ness.

Add IsSpecialFilename, to check if a string is one of the known special
filenames. Use it from GetCanonicalName.

Add IsSpecialFile, which covers more special files, as the successor of
IsBuiltinOrCommandLineFile.

Add a templated IsInSpecialFile, which maps any supported type (tokens,
decls, locations, etc) to a file, and then checks if said file is
special.

I don't see this solving any concrete bugs, but it seems like an
eaiser-to-use and more complete model.