Add and use IsNamedFile and IsInNamedFile
The IsBuiltinFile and IsBuiltinOrCommandLineFile predicates were used in
a few places to determine if includers/includees were actual files.

This would fail to match some non-files like "<stdin>", but was also not
very ergonomic, as callers would have to resolve file entry in order to
check for filey-ness.

Add IsNamedFile as the logical opposite of IsBuiltinOrCommandLineFile,
but that covers even more special files.

Add a templated IsInNamedFile, which maps any supported type (tokens,
decls, locations, etc) to a file, and then checks if said file is
named.

I don't see this solving any concrete bugs, but it seems like a more
consistent model, and the building blocks are more reusable.