Determine provided alias template types correctly
The problem solved here is with alias templates with template
specialization underlying type, like this:

template <typename T>
using Alias = Tpl<T>;

Because the underlying type written in the declaration is `Tpl<T>`,
a final specialization type like `Tpl<int>` is never considered to be
provided if analyzing only `Tpl<T>`. Hence, the concrete underlying type
should be passed for the analysis in `GetProvidedTypesForAlias`.