Report macro-in-macro uses immediately
We used to delay analysis of macro references inside macro definitions
until preprocessing was completely done. There doesn't appear to be any
reason to delay reporting; we can rely on the preprocessor to answer
whether a nested macro is defined at point-of-use or not.

This removes more redundant state tracking and prepares for stronger
macro analysis.