Discern function-like vs. object-like macros in macro bodies
We used to analyze macro definitions to the best of our ability, looking
for identifier tokens that will lead to nested macro expansion. These
are then reported as uses, so dependencies between macros are reflected
in the include dependencies (which isn't always the right call -- some
macros really want to delay the dependency until expansion, but I
digress).

Our analysis was a little too shallow, however, and we couldn't tell the
difference between object-like macros (SYMBOL) and function-like
macros (EXPR(x)), which led to strange conflicts where Clang would
understand that an identifier inside a macro could never expand, but
IWYU would think it could, and push for a macro definition #include.

The case in point was:

    #define BAR top::assert::bar()

where 'assert' looks dangerously like the macro in <assert.h>. But
'assert' is guaranteed by the standard to be function-like, and can
never expand in that context.

Improve the analysis to emulate Clang's rules for macro expansion
viability, and in particular tell the difference between object-like and
function-like macros.

(It would obviously be cool to be able to analyze macro bodies as if
they were 'real code' [think templates], but there's no principled way
to reason about a macro prior to expansion, so that's still written in
the stars.)

Fixes issue #1257.