Fix detection of substituted tpl param in alias
It turns out that the presence of `SubstTemplateTypeParmType` inside
an alias type sugar chain doesn't necessarily mean that it is
a user-provided template parameter. It may occur e. g. when an alias
template is used inside the `typedef` declaration, and its template
parameter is specified inside that declaration. The regression had been
introduced in 2d26dd1b33 and found on MS STL `vector` implementation
when IWYU started to require `_Simple_types` type info for such a code:

std::vector v{1, 2, 3};
(void)v.begin();

due to the returned `iterator` type alias.