Fix location info on instantiated alias tpl scan
Prior to this, template specializations which are used in alias template
definition and actually provided by it might be misattributed
to alias template instantiation locations. Currently, however, to have
this effect, such template specialization types should be explicitly
specified in alias template arguments.

It may look like this improvement is of marginal use. However, it may
become more useful if InstantiatedTemplateVisitor starts to report types
and/or templates not specified in template arguments.