[ci] Disable CMake RPATH stripping
cmake --install by default runs a step to clear RPATH from executables.

Since we install IWYU into the build directory itself, no files are
moved around, but the RPATH settings the linker put in to point to our
LLVM/Clang dylib dependencies will be deleted.

There's some more detailed information on this here:
https://gitlab.kitware.com/cmake/community/-/wikis/doc/cmake/RPATH-handling

This causes the dogfood step to fail, as it doesn't know where to look
for dylibs.

It's not clear why this started failing now, maybe some config changed
in the apt packaging or in LLVM CMake modules.

Explicitly tell CMake to preserve the link RPATH.