Simplify non-member function definition handling
Function definitions can only be reported from VisitFunctionDecl, so
there's no need to search the AST for every use report in
ComputeUseFlags.

(The same is not true for IsNodeInsideCXXMethodBody; that will actually
be true for arbitrarily deep subtrees.)

Instead, pass the use-flag explicitly in VisitFunctionDecl when
reporting declarations.

While here, clarify a comment and make more const-correct.