Decouple handling of fn references from fn calls
Function references differ from calls in that the return type should not
be reported in the former case. It was achieved by means of (not fully)
complemetary `if`-conditions inside `TraverseDeclRefExpr` and
`IwyuBaseAstVisitor::HandleFunctionCall` functions stating that fn refs
should be analyzed only for instantiated template functions with
definitions, meanwhile return types should be reported in the opposite
case. But it is not fully correct. In fact, the full return type info is
required on any function call. It is just convenient to report it on
instantiated template visitation because the type may be provided by
some `typedef`-ed template argument. But it stops working e.g. when
the function template has no definition in the translation unit, just
a declaration. This change enables separate improvement of function
reference and function call handling.

No functional change is intended.