Fix handling function references
Additional filtering of function references has been removed. It is not
needed now because reporting of return type is avoided.

This fixes reporting of template argument type when the template has no
definition in the translation unit but the full type info is needed
to build the instantiated function prototype, as well as handling
instantiations of regular member functions of class templates. It was
a regression introduced in #1300.

`nodes_to_ignore_` restored after function traversal so as to avoid
short-circuiting inside `IwyuBaseAstVisitor::HandleFunctionCall` which
could occur because function declaration was already traversed on its
`DeclRefExpr` handling.