Keep directly included overloads
I've decided to report only that overloaded function decls which are
already directly `#include`d, as opposed to the suggestion
in the removed TODO, so as to avoid lots of unwanted `#include`s
in real-world STL cases (things like `std::begin` may be defined in many
headers). When no overload is included directly but the name is
qualified (i.e. can not be found via ADL), an arbitrary (the first one)
overloaded declaration is suggested to keep the code compilable.

Filtering of suitable overloads has not been done.

This fixes #636, #823, #1111, #1293, and, partially, #1250.