Fix provision analysis with public headers
This fixes determination of types provided with type aliases and
function declarations (fn return, "autocast") in the presence of mapping
files. The problem was in that the `PublicHeaderIntendsToProvide`
function considers as provided all the headers transitively included
from a public header, despite they are may be themselves public and not
have the explicit public-to-public mapping specified. It is convenient
to avoid noisy suggestions of the library internals when handling
instantiations of some external library templates, but may hinder
expressing author intent when writing e.g. some own (user's) `typedef`
(non-)providing a type from such a library.

To achieve the more correct behavior, the logic from
`OneUse::SetPublicHeaders()` function has been reused which takes into
account only explicitly specified mappings. The drawback is that the use
site vs. the decl site difference for default template arguments becomes
worse (see `TestDefaultTplArgs()`). The use site handling still uses
`PublicHeaderIntendsToProvide` (which is common way for template
instantiation handling, as stated before). It may end up in that IWYU
may suggest to remove an indirectly providing header from the template
defining file (`-d1.h` from `-def_tpl_arg.h` in the test case)
not suggesting to `#include` the default argument header at the use site
because at the time of analysis `-d1.h` is present in `-def_tpl_arg.h`
and "provides" all its directly and indirectly included stuff. OTOH, it
may be fixed easily by means of including the appropriate header
in the template instantiating file manually, and the IWYU behavior
should become consistent.

The main goal of these changes is to introduce `IsDirectlyIncluded`
function which may be reused when fixing `UnresolvedLookupExpr`-related
bugs (`PublicHeaderIntendsToProvide` doesn't work well in real-world
cases for it).