[ci] Generalize workaround for broken packaging
Every time a new major version of the Clang APT packages rolls out, our
build breaks because libclang-NN.so.1 is not properly installed. Then
after a few weeks it usually sorts itself out, for some reason.

We don't use this shared library, but the LLVM/Clang CMake modules
self-verify and fail if the file does not exist.

We've worked around it for a number of releases by simply touching the
missing filename after the build fails for the first time. But since
this appears to be a recurring problem, generalize to:

* find the latest /usr/lib/llvm-NN directory
* extract the NN major version number
* use the version NN to dynamically generate an .so name to touch

Whether the file exists or not, this should placate CMake.