[clang compat] Add printing policy when generating fwd decls
Clang changed its formatting of decls in this patch:
https://github.com/llvm/llvm-project/commit/9391ff8c86007562d40c240ea082b7c0cbf35947

A tiny bug was introduced, where sometimes DeclPrinter puts a spurious
extra space after the tag specifier (class, struct, etc). We generate
forward declarations using the DeclPrinter, and the resulting format is
pinned down in our tests.

This caused the following regression in IWYU:

    -template <typename T> class FinalTemplate;
    +template <typename T> class  FinalTemplate;

Explicitly use a printing policy with PolishForDeclaration, which omits
semantic attributes (and thereby the extra space) from the output.