Add and use DropFrom string util
We had one place in MungedForwardDeclareLineForTemplates where Split was
used to cut a string on a separator.

Using Split for this is both noisy at the callsite and non-performant.

Add DropFrom, which finds the separator and drops the rest of the string
from there. Use it to simplify MungedForwardDeclareLineForTemplates a
little.