Report better location for BinaryOperator types
The current location for a BinaryOperator points to the actual
operator ('+', '-', etc), not the terms we're reporting types for.

This confused the logic in GetCanonicalUseLocation trying to determine
who is responsible for the use, so it said the macro spelling was
responsible for pointer arithmetic on arguments.

Strip parentheses and implicit casts to get to the actual term
expressions, and explicitly use their location.

Add tests for pointer arithmetic in macros, both on macro args and
entirely encapsulated in the macro.

Fixes #1483.