[iwyu_tool] Default to -j=0 and infer the core count in that case
Make the behavior of the -j argument more similar to other tools which
offer it. Instead of hanging forever when "zero jobs" are requested,
attempt to automatically pick a job count suitable for the current
system.

Also allow -j to be used as a boolean flag, without passing a value. In
that case, the job count is chosen automatically as in j=0. If the -j
argument is not used at all, operation is still single-threaded
(same behavior as before this patch).