Report unresolved refs to variable templates
The change in badinc.cc is because I1_NamespaceTemplateFn shadow decl
is now reported on CallOverloadWithUsingShadowDecl scanning, hence
using-declaration is considered as being used and may not require
`#include` by itself (as for the other using-decls). If its reference
is removed from CallOverloadWithUsingShadowDecl, the suggestion comes
back, as it should be.

This fixes #792, #966, #1335, #1516, and #1558.