Remove brittle offload_openmp driver testcase
This testcase kind-of proved that our driver drops OpenMP/NVPTX jobs,
but it was heavily dependent on the target environment.

I'd only executed it on my own machine, and our CI environment, neither
of which have any OpenMP toolchains installed.

If such a toolchain _is_ installed the failure diagnostics will not
match, and the test will fail.

Since we don't have any control of the environment for these tests, and
no way to isolate them, remove the test.