Use Clang builtin headers in-place
In standalone mode we would copy the Clang builtin headers by
replicating one of Clang's targets (clang-resource-headers), poorly.

Instead, build a -resource-dir argument implicitly on IWYU startup to
point directly to the Clang headers. The path to the headers are
computed by the Clang Driver based on the Clang executable path,
resolved by CMake.

This does not need to happen for the external-project build
mode (building as part of the LLVM tree); there the headers are already
available in the build tree, and 'include-what-you-use' is installed in
the same place as 'clang' so the default Clang Driver policy for
resource dir lookup will work.