Take most basic method for author-intent analysis
Because overriding method prototypes need generally neither the full
info nor the forward-declaration for the return type and for
the parameter types, the author intent analysis should be performed only
for the original declaration from the upmost class in the hierarchy.
This change does that at the call site.

This is a step towards fixing #105.

The test case has been partially taken from #675.