Use HeaderSearch for include name lookup instead of @file
The @headername directive is a Doxygen alias used in libstdc++, but
other projects have also picked up the convention. When generating
implicit mappings for @headername, we used to use the Doxygen @file
directive to get the from-name and @headername to get the preferred
to-names.

There were a couple of problems with our interpretation of these
pragmas:

* @file is manually maintained, and may be empty, incorrect or ambiguous
* Doxygen requires that @file contains enough of the file path in the
  build tree to be unique, so the name spelled out there is not
  necessarily the include-name we should use (e.g. many files are
  prefixed with "include/" for disambiguation)
* We used to assume all @headername names were system headers. This is
  true for libstdc++, but not necessarily for other projects.

Change this to use the private header FileEntry instead, and use the
Clang preprocessor to get the actual include-name that brought the file
into the source manager in the first place.

Move the include-name handling to after a @headername has been found to
avoid the work if there's no mapping information anyway.

Change quoting for the to-name so it's not hard-coded to angle quotes,
but rather use the same quote style as the private header (figuring out
the quote style for an arbitrary public header string is non-trivial).
Update a test to use the right quoting.

This fixes a bunch of mis-mappings for libstdc++.