Consider associated header desired includes
The statement that IWYU doesn't change associated headers is clearly
wrong. When handling a main file, it should take into account what it
suggests to do with its associated headers, not what those headers look
like at the moment of the analysis.

It is assumed that the desired `#include`s of the associated headers
have been calculated before handling the main file, as it was already
assumed in the code.

This fixes #1142.