[ci] Add more systemic workaround for broken packaging
Add a script to disable target-vs-file wholesale in LLVM and Clang CMake
export modules.

We do this more heavy-handed disabling because our CI is blocked if
llvm-toolchain is in a bad state, even for modules we don't need, and we
can't verify compat with upstream or get patches tested.

This should fix the build so we can rely on our CI again.