Improve libstdc++ private-to-private mappings
The mapping generator for libstdc++ generates private-to-private mappings, to
form a connected graph from the most private header to the most public, via any
number of intermediate private headers.

Unfortunately I mistakenly skipped all private headers for which there was
already a private-to-public mapping. I think the intent was to avoid visibility
mismatches, i.e. make sure we didn't create a new mapping for <x> with
visibility private, where there was already one with visibility public.

That check, however, does not solve that problem. The keys of public_mappings
are private headers, and they will be marked visibility=private in the resulting
mapping records. The potential conflict would be if we mapped a private header
<x> to a private header <y> where <y> had previously been marked as public.

As it turns out, I don't think that can ever happen, because the private header
and public header sets are disjoint.

Add an assertion to prove they're disjoint. Remove the broken skip.

Regenerate both gcc.stl.headers.imp and built-in libstdc++ mappings.