[fix_includes] Use raw strings to hush SyntaxWarnings in tests
Python3.12 upgraded a deprecation warning for invalid escape sequences to a
syntax warning, which now fires in a couple of places in
fix_includes_test.py:

    fix_includes_test.py:141: SyntaxWarning: invalid escape sequence '\S'
      filenames = re.findall('^(\S+) should [...]

Use raw strings where necessary to silence this. Unfortunately we also have an
unescaped "\s" in a docstring, so we have to live with a very jarring r""" start
token there.