Don't attempt to clear elaborated type duplicates
It was assumed in #1232 that at most one elaborated type can "own" a tag
declaration (i.e. none of possible following similar types redeclares
that declaration), hence there should be no duplicates in fwd_decl_map.
Even if that assumption is true, an ElaboratedTypeLoc may be handled
more than once if it corresponds to different AST nodes. It happens when
that type is used in a declaration of several variables.

Elaborated types should obviously not be removed in any case because
they are part of user's main code and not of #include or fwd-decl lists.

This fixes #1674.