Make "assignable" trait handling more fine-grained
This patch is not consistent in what should be done with _possible_
conversions. That means: if the types don't satisfy the trait at
the time of analysis, should they be nevertheless complete to prevent
a UB if some conversion operators appear in the future? It has been
generally answered "yes" for this patch, but not for the pointer
assignments. The cases when a type starts inheriting another type,
previously unrelated, seem to be rare (or not so rare?). However, this
has not been applied to reference assignments, when the full rhs class
type is always required. This is because I'm really not sure which
approach should be preferred... But I think IWYU hardly can (and should)
prevent any possible UB after user's code future refactoring, especially
in cases like described in #745, otherwise it might require completely
unrelated types.

Nevertheless, this fixes #1615.