Suppress ambiguous private-to-public libstdc++ mappings
If a private header maps to more than one public header, it's often the case
that it's just an internal utility, and users will never see symbols exported
from the private header.

We generate implicit private-to-public mappings by building a reverse map from
a private header to all public headers that include it. Add a rule to suppress
any private-to-public mappings with more than one public header.

Also add an explicit override for explicitly mapping a private header to a
public in the case where a private header is used by multiple public, but only
exported from one of them.

Regenerate all libstdc++ mappings with:

    mapgen/iwyu-mapgen-libstdcxx.py \
        --lang=c++ \
        /usr/include/c++/11 \
        /usr/include/x86_64-linux-gnu/c++/11 \
        > libstdcxx-11.cc

    mapgen/iwyu-mapgen-libstdcxx.py \
        --lang=imp \
        /usr/include/c++/11 \
        /usr/include/x86_64-linux-gnu/c++/11 \
        > gcc.stl.headers.imp

Fixes #1690.