Minor fixes to man page
For the experimental flags, we distinguish between (bold) terminals,
italic non-terminals, and regular syntax. We also refer to the non-
terminal in the text.

In the section of mapping files, we move the quotes around "visibility"
up to the syntax summary line like quotes around the other entities.
This seems more consistent.