Fix sys/ttycom.h mapping
Some BSDs define winsize in sys/ttycom.h, so it was originally mapped to
termios.h. A mapping for the winsize symbol has since been added and the
remaining symbols should be imported via sys/ioctl.h. While here, group
it with other system headers that are not present on Linux.