Improve "convertible" trait handling
Because '__is_convertible_to' is a synonym for '__is_convertible', and
already has some test coverage, test cases for it have not been added.
OTOH, '__is_nothrow_convertible' has also been handled similarly
in IWYU, so the test cases for it may be redundant as well...

The change in the 'ptr_ref_aliases' test is because the only
'IndirectBase' use that was considered as a full use by IWYU was inside
'__is_convertible_to' traits.