Add all reproducible bugs to bugs suite
Now that the bugs suite exists to do negative regression testing
(i.e. notice when bugs are fixed), add repro cases for all bugs where
I've been able to produce a fully contained repro, or one that only
depends on libc, libstdc++ or libc++.

So far these are probably only reproducible in a Linux environment with
GCC and Clang available.

Once this is merged, our CI environment will fail the build if a PR
fixes a bug -- not that there's anything wrong with that, but we need to
make sure it comes with fully reduced regression tests. Once that's
taken care of, we can remove the bugs test and close the bug.

Roughly 60 bugs are getting test coverage with this patch.