Always use "use location" as canonical use location for a decl
Currently, the headers that IWYU suggests to add to a header file
defining a macro depend on whether that macro is used in the corresponding
implementation file that causes the header file to be analyzed. So if
a file a.h defines a macro that uses various declarations and the corresponding
file a.c does not use that macro, then IWYU will suggest removing any
headers that provide those decls from a.h. On the other hand, if a.c
does use the macro, then IWYU will suggest keeping the headers.

This means that if a.c does not use the macro, then any other files that
include a.h and use the macro will end up with missing header files as
the required header files to use the macro will not be added to a.h and
will not be added to the file using the macro either, causing compilation
failures.

To fix this problem, let's always use the "use location" for decls within
a macro as the canonical location, so that the required headers are always
added to the file that's expanding the macro.

Fixes #1737