Add array "constructible" trait handling
Since C++20 (apparently after P0960), "constructible" traits may
evaluate to true when testing array construction. This changeset adds
reporting of types on which the result depends in such cases.

As for the case of ctors with multiple args handling of which was added
earlier, firstly the argument types are checked whether they might be
converted to the array element type. For example:
__is_constructible(int[2], Class&, int*)
is always false regardless of whether Class has a conversion function
to int because int* cannot be implicitly converted to int, hence
the complete Class type is not needed. So as not to write all the cases
by hand, IsConvertible function has been introduced, the logic of which
has been taken from the clang source. Some of its code
(EnterExpressionEvaluationContext, sfinae_trap.hasErrorOccurred() check)
is not covered by tests and has been added only for generality. It seems
that it makes sense only with record types for which this function is
not used currently. OTOH, ContextRAII has not been used because Sema
should already be in the global (translation unit) context at the time
of the IWYU analysis.