Handle array references in type traits
Due to array to pointer implicit conversion, arrays may be used where
pointers to their elements are acceptable. Direct or typedefed array
types in type traits should currently be reported unconditionally,
because `CanForwardDeclareType` returns false for them. This change adds
handling of arrays under reference types.

`ASTContext::getAsArrayType` desugars the given type to the array type
and merges the cv-qualifier with the one of the array element type.

One comment has been corrected because `ReportTypeUse` performs
reporting not only for record types but also for elements of arrays
of them.