Refactor some functions to work with QualType
IsDerivedToBasePtrConvertible should be made working with arrays because
they may be considered like pointers to their element types due to
array-to-pointer implicit conversion. In order to determine
the cv-qualifier of the resulting pointer correctly, the function should
take QualType, because a cv-qualifier applied to an array type
propagates to its element type.

base_ptr_type parameter could in principle be kept as 'const Type*', but
I've decided to change its type too for symmetry.

RemoveReference function has been transformed to work with QualType so
that IsDerivedToBasePtrConvertible can be supplied with its result.

No functional change.