Add bugs suite skeleton
Add a README to explain the purpose and some procedure.

Add a new --extra-suite argument so we can dynamically register tests/
subdirectories as suites from the command-line.

Set up a build step to run the bugs suite from CI.

Add an exemplary test for an old issue so the suite can be executed.