Rename built-in mappings to internal mappings
The word "built-in" is important in compiler design; it usually refers
to an intrinsic function whose implementation is generated dynamically
by the compiler (e.g. __builtin_addcb or pow from math.h).

IWYU has historically distinguished between two different kinds of
mappings, built-in and external, where the built-in mappings are
hard-coded in include_picker.cc and the external mappings are declared
in YAML and parsed on demand.

The use of the word "built-in" here is confusing, particularly since we
carry mappings for various intrinsics headers and symbols.

Rename all mentions of "built-in" or "builtin" mappings to use
"internal" instead.