Handle __builtin_types_compatible_p
__builtin_types_compatible_p is a C language trait that was almost like
"is same" before C23. However, N3037 allowed different tag type to be
compatible in the same translation unit provided the tags are identical
and the contents are similar. Types with identical tags may be different
when they are in different scopes. The interesting for IWYU case is when
one declaration is in the file scope and the other is in a block scope.
The file scope declaration may be in a header file and may be required
to correctly determine type compatibility.

The types to report may be buried under an arbitrary number of derived
types like pointers or function types. Moreover, field types of structs
and unions should be recursively handled as well. For example:

struct Inner { int i; };
struct Outer { struct Inner** p; };
extern struct Outer o;

int Fn() {
  struct Inner { int i; };
  struct Outer { struct Inner** p; };
  return __builtin_types_compatible_p(struct Outer, typeof(o));
}

Not only Outer but also Inner should be reported at the trait location
for the trait to be evaluated correctly. OTOH, if Outer structures
contained another fields which differ in type or name, reporting Inner
would better to be avoided because the trait is guaranteed to be 0.
The only way to achieve this is to (almost) evaluate the trait inside
IWYU.