Fix opaque declarations of elaborated enums
C and C++ don't allow elaborated types to introduce new enumerations,
so opaque declarations of enumerations with fixed underlying types are
still needed. If the underlying type is not fixed, the full definition
is required, but in this case CanForwardDeclareType returns 'false'.

Prior to this, for elaborated types naming enumerations with fixed
underlying type, IWYU required neither the full definition nor an opaque
declaration.