Do not suggest forward-decls for implicit code
There is no need to change VisitUsingType because using-types inside
implicit code are already skipped in TraverseType and TraverseTypeLoc.

Another way to fix it might be to move ReportDeclForwardDeclareUse calls
into Visit...TypeLoc functions which are not called for implicit code
(due to IwyuAstConsumer::TraverseTypeLoc logic). But this breaks fwd-
declaring of types from dynamic exception specifications because they
are `Type`s in the AST and not `TypeLoc`s (there is a TODO inside
RecursiveASTVisitor).

This fixes #1774.