Add symbol mapping for std::nullptr_t
std::nullptr_t in libstdc++ is defined in the private, arch-specific
header bits/c++config.h. That header has a @headername directive
pointing to <version>, which is only defined for C++20, and doesn't
really provide std::nullptr_t.

I can't think of a principled way to fix this, so add a simple libstdc++
symbol mapping for std::nullptr_t to <cstddef>.

Drive-by improvement of the comment for std::declval for consistency.

Fixes #1791.